refactor(dev): split up build and watch into separate modules
also, split out compile failure logging into its own module
